backend_unified/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
requests>=2.31.0
pymupdf>=1.23.0
groq>=0.4.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
